+++ docs/changelog.md
@@ -63,6 +63,10 @@
 
 ## 更新说明详情
 
+1.11.36 add: add stock_fear_greed_funddb interface
+
+    1. 新增 stock_fear_greed_funddb 接口
+
 1.11.35 fix: fix stock_board_concept_hist_ths interface
 
     1. 修复 stock_board_concept_hist_ths 接口
@@ -2971,6 +2975,8 @@
     3. 修改 air_quality_hist 接口的请求日期格式，从 '2022-01-01' 为 '20220101'
 
 ## 版本更新说明
+
+1.11.36 add: add stock_fear_greed_funddb interface
 
 1.11.35 fix: fix stock_board_concept_hist_ths interface